Add store code logic
Add store code logic. Update method in tests, that asserts validation
errors. Update validation rules in store lottery request. Create and
update tests.

+++ app/Http/Requests/Api/V1/StoreLotteryRequest.php
@@ -34,20 +34,20 @@ class StoreLotteryRequest extends FormRequest
             'numbers_count' => [
                 'bail',
                 'required',
-                'numeric',
+                'integer',
                 'min:1',
                 'max:255',
             ],
             'numbers_from' => [
                 'bail',
                 'required',
-                'numeric',
+                'integer',
                 'lte:numbers_to',
             ],
             'numbers_to' => [
                 'bail',
                 'required',
-                'numeric',
+                'integer',
                 'gte:numbers_from',
             ],
         ];
